requirements.txt
# Core dependencies
Pillow>=9.0.0
numpy>=1.21.0

# Development dependencies (optional)
# Install with: pip install -r requirements.txt -r requirements-dev.txt
